Fix: Disable ESLint and Typescript during production builds
This commit updates the next.config.ts file to disable ESLint and Typescript checks during production builds, in addition to development and CI builds. This is because we run separate linting and type checking processes in CI, so we don't need to run them during the build process.

Additionally, this commit updates the tsconfig.ci.json file to include new API routes related to contracts. This ensures that these routes are properly type-checked during CI builds.