Refactor API authentication and update dependencies
This commit refactors the API authentication logic by introducing a `withAuth` utility function to handle user authentication and profile retrieval in API routes. This change promotes code reusability and improves the overall structure of the API endpoints.

Additionally, the commit updates several dependencies, including eslint, electron-to-chromium, and others, to their latest versions. This ensures that the project is using the most up-to-date and secure versions of its dependencies.

The `layout-debug.tsx` file was removed as it is no longer needed.